pr fix: remove mockContractAndRatesDraftWithDocuments and ensure error message displays only on unlocked submissions on the summary page
