+++ services/app-web/src/testHelpers/apolloMocks/index.ts
@@ -9,7 +9,6 @@ export {
     mockUnlockedHealthPlanPackageWithDocuments,
     mockUnlockedHealthPlanPackageWithOldProtos,
     mockSubmittedHealthPlanPackageWithRevisions,
-    mockContractAndRatesDraftWithDocuments,
     mockUnlockedHealthPlanPackage,
 } from './healthPlanFormDataMock'
